requirements.txt
Flask==2.0.1
Flask-Reuploaded==1.1.0
Flask-Uploads==0.2.1
google-cloud-vision==2.3.1
gunicorn==20.1.0
Werkzeug==2.0.1